% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_plot.R
\name{area_plot}
\alias{area_plot}
\title{Peak area plot}
\usage{
area_plot(rt, int, peak_info)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{peak_info}{a \code{data.frame} of peak info (see \code{\link{find_peaks}})}
}
\description{
Plot the chromatogram with the peak area of all detected peaks, highlighted
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
