% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_153-data.R
\docType{data}
\name{chrom_153}
\alias{chrom_153}
\title{chrom_153}
\format{An object of class \code{list} of length 4.}
\usage{
chrom_153
}
\description{
List of four elements; where each element is a tibble containing retention time (rt) and intensity (int) for a
product transition of the parent 153
}
\keyword{datasets}
