% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_181-data.R
\docType{data}
\name{chrom_181}
\alias{chrom_181}
\title{chrom_181}
\format{An object of class \code{list} of length 4.}
\usage{
chrom_181
}
\description{
List of four elements; where each element is a tibble containing retention time (rt) and intensity (int) for a
product transition of the parent 181
}
\keyword{datasets}
