% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_194-data.R
\docType{data}
\name{chrom_194}
\alias{chrom_194}
\title{chrom_194}
\format{An object of class \code{list} of length 4.}
\usage{
chrom_194
}
\description{
List of four elements; where each element is a tibble containing retention time (rt) and intensity (int) for a
product transition of the parent 194
}
\keyword{datasets}
