% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_plot.R
\name{chrom_plot}
\alias{chrom_plot}
\title{Chrom plot}
\usage{
chrom_plot(rt, int, peak_info = NULL)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{peak_info}{an optional  \code{data.frame} of peak info (see \code{\link{pick_peaks}}) (default is\code{NULL})}
}
\description{
Plot the chromatogram. If a \code{peak_info} \code{data.frame} is supplied, then the apex's of all detected peaks are indicated
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
