% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Peak Detection}
\usage{
find_peaks(rt, int)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}
}
\value{
a \code{peak_info} \code{data.frame}
}
\description{
Detect peals using \code{pracma::findPeaks}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
