% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chrom_max.R
\name{get_chrom_max}
\alias{get_chrom_max}
\title{Find Chromatographic Peak Apex}
\usage{
get_chrom_max(rt, int)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}
}
\value{
a numeric value for the \code{rt} at the apex with the maximum intensity
}
\description{
Find Chromatographic Peak Apex
}
