% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_info.R
\name{get_peak_info}
\alias{get_peak_info}
\title{Peak info}
\usage{
get_peak_info(rt, int, peak_limits)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity#' @param idx}

\item{peak_limits}{a \code{data.frame} indicating the left, right and apex indices for all detected peaks}
}
\value{
a \code{peak_info} \code{data.frame}
}
\description{
Create a \code{peak_info} \code{data.frame}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
