% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Peak Detection}
\usage{
get_peaks(rt, int)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}
}
\value{
a \code{peak_info} \code{data.frame}. See also \link{\code{get_peak_info}}
}
\description{
Detect peals using \code{pracma::findPeaks}
}
