% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{merge_peaks}
\alias{merge_peaks}
\title{Merge peaks}
\usage{
merge_peaks(rt, int, idx, peak_info)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity#' @param idx}

\item{idx}{a numeric vector of peak id's to merge.}

\item{peak_info}{a \code{data.frame} of peak info (see \code{\link{find_peaks}})}
}
\value{
an updated \code{peak_info} \code{data.frame}
}
\description{
Merge multipe detected peaks toegther into a single chromatographic feature
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
