% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{peak_detection}
\alias{peak_detection}
\title{Peak Detection}
\usage{
peak_detection(rt, int)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}
}
\value{
a \code{peak_info} \code{data.frame}
}
\description{
Detect peals using \code{pracma::findPeaks}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
\keyword{internal}
