% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_}
\alias{pick_}
\title{Run pick function by group}
\usage{
pick_(unit, time, picker, ...)
}
\arguments{
\item{unit}{vector indicating group}

\item{time}{time vector}

\item{picker}{a single-unit pick-function}

\item{...}{arguments passed to picker}

\item{per_unit}{list of arguments passed to picker (like MoreArgs in mapply)}
}
\value{
ragged list of per-unit indexes
}
\description{
Run pick function by group
}
