% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_first}
\alias{pick_first}
\title{Pick observations occurring at the first-recorded time.}
\usage{
pick_first(time)
}
\arguments{
\item{time}{time vector}
}
\value{
index
}
\description{
Pick observations occurring at the first-recorded time.
}
