% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_lag_heads}
\alias{pick_lag_heads}
\title{Pick observations occurring at time pivot.}
\usage{
pick_lag_heads(time, pivot)
}
\arguments{
\item{time}{time vector}

\item{pivot}{pivot time point}
}
\value{
index
}
\description{
Pick observations occurring at time pivot.
}
