% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_lag_tails}
\alias{pick_lag_tails}
\title{Pick observations ocurring prior to head observations
and towards the pivot.  Ignore tails equal to pivot.}
\usage{
pick_lag_tails(time, pivot)
}
\arguments{
\item{time}{time vector}

\item{pivot}{pivot time point}
}
\value{
index
}
\description{
Pick observations ocurring prior to head observations
and towards the pivot.  Ignore tails equal to pivot.
}
