% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_last}
\alias{pick_last}
\title{Pick observations occurring at the last-recorded time.}
\usage{
pick_last(time)
}
\arguments{
\item{time}{time vector}
}
\value{
index
}
\description{
Pick observations occurring at the last-recorded time.
}
