% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_median}
\alias{pick_median}
\title{Pick observations occurring nearest the median time.}
\usage{
pick_median(time, smaller = TRUE)
}
\arguments{
\item{time}{time vector}

\item{smaller}{if TRUE (default) return the smaller of two central 
values, otherwise return larger}
}
\value{
index
}
\description{
Pick observations occurring nearest the median time.
}
