% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_peaks.R
\name{pick_peaks}
\alias{pick_peaks}
\title{Pick peaks}
\usage{
pick_peaks(rt, int, target_rt, rt_window = 20)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{target_rt}{a numeric value of a target retention time for targeted peak picking. Retention time must be in the format
M.SS (ie, 5.48)}

\item{rt_window}{a numeric value (in seconds) for retention time window. Any peaks detected within the specified \code{rt_window} of the
\code{target_rt} and combined into a single chromatographic feature. This is mainly of use for when poor quality peaks exhibit splitting,
or intense shoulder peaks}
}
\value{
a \code{peak_info} \code{data.frame}
}
\description{
Pick peaks
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
