% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_pivot_distance}
\alias{pick_pivot_distance}
\title{Pick observation order based on time (distance 
to pivot).}
\usage{
pick_pivot_distance(time, pivot)
}
\arguments{
\item{time}{time vector}

\item{pivot}{pivot time point}
}
\value{
index
}
\description{
Pick observation order based on time (distance 
to pivot).
}
