% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_prior}
\alias{pick_prior}
\title{Pick observationsn occurring prior to a pivot time point}
\usage{
pick_prior(time, pivot)
}
\arguments{
\item{time}{time vector}

\item{pivot}{pivot time point}
}
\value{
index
}
\description{
Pick observationsn occurring prior to a pivot time point
}
