% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pick_time}
\alias{pick_time}
\title{Pick observations occurring at a time point}
\usage{
pick_time(time, t)
}
\arguments{
\item{time}{time vector}

\item{pivot}{pivot time point}
}
\value{
index
}
\description{
Pick observations occurring at a time point
}
