% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{picker_proxy}
\alias{picker_proxy}
\title{Create a picker proxy object}
\usage{
picker_proxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
deck to modify}

\item{session}{the \code{Shiny} session object to which the picker widget belongs;
usually the default value will suffice.}
}
\description{
Creates a picker-like object that can be used to update a picker object that has already been rendered.
}
