% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area.R
\name{plot_area}
\alias{plot_area}
\title{Plot chromatogram with peak area}
\usage{
plot_area(rt, int, peak_info)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{peak_info}{a \code{data.frame} of peak info (see \code{\link{get_peaks}})}
}
\value{
a \code{ggplot} object
}
\description{
Plot the chromatogram with the peak area of all detected peaks highlighted
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
