% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chrom.R
\name{plot_chrom}
\alias{plot_chrom}
\title{Plot the chromatogram.}
\usage{
plot_chrom(rt, int, peak_info)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{peak_info}{an optional  \code{data.frame} of peak info (see \code{\link{get_peak_info}}) (default is\code{NULL})}
}
\description{
If a \code{peak_info} \code{data.frame} is supplied, then the apex's of all detected peaks are indicated
}
