% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_peak.R
\name{smooth_peak}
\alias{smooth_peak}
\title{Smooth Peak}
\usage{
smooth_peak(rt, int, ford)
}
\arguments{
\item{rt}{a numeric vector of retention time}

\item{int}{a numeric vector of intensity}

\item{ford}{a numeric value for the filter order (2 = quadratic filter, 4 = quartic)}
}
\value{
a numeric vector of smoothed intensity
}
\description{
Use an optimised Savitzky-Golay function to smooth a chromatogram
}
