% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{update_picker}
\alias{update_picker}
\title{Send commands to a picker instance in a \code{Shiny} app}
\usage{
update_picker(proxy, view_state)
}
\arguments{
\item{proxy}{picker proxy object}

\item{...}{unused}
}
\description{
Send commands to a picker instance in a \code{Shiny} app
}
\seealso{
\link{picker_proxy}
}
