% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picker.R
\name{picker}
\alias{picker}
\title{Title}
\usage{
picker(
  coords,
  colors,
  labels,
  label_coords = NULL,
  polygons = NULL,
  point_color_polygons = NULL,
  show_controls = TRUE,
  scatter_props = NULL,
  deck_props = NULL,
  text_props = NULL,
  polygon_props = NULL,
  xrange = NULL,
  yrange = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{coords}{data.frame with two columns. First has x, second has y coordinates.}

\item{colors}{vector of hex colors, one for each row of \code{coords}.}

\item{labels}{vector of point labels used for tooltips on hover.}

\item{label_coords}{data.frame with three columns 'x', 'y', and 'label'. Used for text layer.}

\item{show_controls}{Should control panel be shown? Default is \code{TRUE}.}

\item{scatter_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/layers/scatterplot-layer}{ScatterplotLayer}.}

\item{deck_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/core/deck}{Deck} instance.}

\item{text_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/layers/text-layer}{TextLayer}.}

\item{width}{width of htmlwidget.}

\item{height}{height of htmlwidget.}

\item{elementId}{id of htmlwidget.}
}
\value{
renders html widget
}
\description{
Title
}
