% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picker.R
\name{picker}
\alias{picker}
\title{Render a Picker Widget}
\usage{
picker(
  coords,
  colors,
  labels,
  title = NULL,
  label_coords = NULL,
  polygons = NULL,
  point_color_polygons = NULL,
  show_controls = TRUE,
  grid_legend_items = NULL,
  scatter_props = NULL,
  deck_props = NULL,
  text_props = NULL,
  polygon_props = NULL,
  xrange = NULL,
  yrange = NULL,
  xaxs = 0.04,
  yaxs = 0.04,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{coords}{data.frame with two columns. First has x, second has y coordinates.}

\item{colors}{vector of hex colors, one for each row of \code{coords}.}

\item{labels}{vector of point labels used for tooltips on hover.}

\item{label_coords}{data.frame with three columns 'x', 'y', and 'label'. Used for text layer.}

\item{polygons}{data.frame containing at minimum columns 'x1', 'x2', 'y1', 'y2',
that define the polygons to draw and 'color' that defines the color.}

\item{point_color_polygons}{character, a color to make points when polygons are shown e.g. \code{'white'}.}

\item{show_controls}{Should control panel be shown? Default is \code{TRUE}.}

\item{scatter_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/layers/scatterplot-layer}{ScatterplotLayer}.}

\item{deck_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/core/deck}{Deck} instance.}

\item{text_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/layers/text-layer}{TextLayer}.}

\item{polygon_props}{Props passed to deck.gl \href{https://deck.gl/docs/api-reference/layers/polygon-layer}{PolygonLayer}.}

\item{xrange}{range of x-values. Default is \code{range(coords[,1])}.}

\item{yrange}{range of y-values. Default is \code{range(coords[,2])}.}

\item{xaxs}{the fraction to extend \code{xrange} on either side. Default is 0.04.}

\item{yaxs}{the fraction to extend \code{yrange} on either side. Default is 0.04.}

\item{width}{width of htmlwidget.}

\item{height}{height of htmlwidget.}

\item{elementId}{id of htmlwidget.}
}
\value{
renders html widget
}
\description{
Render a Picker Widget
}
